+++ .eslintignore
@@ -1,0 +1,5 @@
+public/sw.js
+public/sw.js.map
+public/workbox-*.js
+public/workbox-*.js.map
+
