+++ .prettierignore
@@ -3,4 +3,11 @@ next-env.d.ts
 node_modules
 yarn.lock
 package-lock.json
-public
+
+
+# Service worker
+public/sw.js
+public/sw.js.map
+public/workbox-*.js
+public/workbox-*.js.map
+
