+++ components/AppFooter/index.ts
@@ -1,0 +1,1 @@
+export { AppFooter } from './AppFooter';
