+++ modules/layout/AppLayout.tsx
@@ -1,6 +1,6 @@
 import { SkipToContent } from 'components/SkipToContent';
 import { Header } from 'modules/mainHeader/Header';
-import { SiteFooter } from 'components/SiteFooter';
+import { AppFooter } from 'components/AppFooter';
 
 type AppLayoutProps = {
   children: React.ReactNode;
@@ -14,6 +14,6 @@ export const AppLayout: React.FC<AppLayoutProps> = ({ children }) => (
 
     <div className="app-container">{children}</div>
 
-    <SiteFooter />
+    <AppFooter />
   </>
 );
